docs: 📄 Complete API > Utility functions
