+++ lib/functions/assert.ts
@@ -1,0 +1,5 @@
+export function assert(condition: any, message: string = "no additional info provided"): asserts condition {
+  if (!condition) {
+    throw new Error("Assertion Error: " + message);
+  }
+}
