+++ lib/functions/create-factory-with-constraint.ts
@@ -1,0 +1,4 @@
+export const createFactoryWithConstraint =
+  <Constraint>() =>
+  <Value extends Constraint>(value: Value): Value =>
+    value;
