+++ lib/functions/is-exact.ts
@@ -1,0 +1,7 @@
+import { Exact } from "../exact";
+
+export const isExact =
+  <Expected>() =>
+  <Type>(value: Exact<Type, Expected>): Expected => {
+    return value;
+  };
