+++ lib/functions/noop.ts
@@ -1,0 +1,1 @@
+export function noop(..._args: unknown[]): void {}
