+++ lib/functions/unreachable-case-error.ts
@@ -1,0 +1,5 @@
+export class UnreachableCaseError extends Error {
+  constructor(value: never) {
+    super(`Unreachable case: ${value}`);
+  }
+}
